refactor: update DeleteProjectUseCase to use direct calls
