+++ src/main/kotlin/di/UseCaseModule.kt
@@ -2,7 +2,7 @@ package di
 
 import logic.useCase.CreateTaskUseCase
 import org.example.logic.useCase.*
-import org.example.logic.useCase.deleteProject.DeleteProjectUseCase
+import org.example.logic.useCase.DeleteProjectUseCase
 import org.example.logic.useCase.deleteTask.DeleteTaskUseCase
 import org.example.logic.useCase.updateProject.UpdateProjectUseCase
 import org.example.logic.useCase.updateTask.UpdateTaskUseCase
