+++ src/main/kotlin/presentation/screens/ProjectsOverviewUI.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.runBlocking
 import org.example.logic.models.AuditLogEntityType
 import org.example.logic.models.Project
 import org.example.logic.useCase.*
-import org.example.logic.useCase.deleteProject.DeleteProjectUseCase
+import org.example.logic.useCase.DeleteProjectUseCase
 import org.example.logic.useCase.updateProject.UpdateProjectUseCase
 import org.example.presentation.role.ProjectScreensOptions
 import org.koin.java.KoinJavaComponent.getKoin
