+++ src/test/kotlin/presentation/screens/ProjectsOverviewUITest.kt
@@ -5,7 +5,7 @@ import org.example.logic.models.*
 import org.example.logic.useCase.GetAllProjectsUseCase
 import org.example.logic.useCase.GetEntityAuditLogsUseCase
 import org.example.logic.useCase.GetProjectByIdUseCase
-import org.example.logic.useCase.deleteProject.DeleteProjectUseCase
+import org.example.logic.useCase.DeleteProjectUseCase
 import org.example.logic.useCase.updateProject.UpdateProjectUseCase
 import org.example.presentation.role.ProjectScreensOptions
 import org.example.presentation.screens.ProjectsOverviewUI
